imapd.c: imapoptions: implement idle timeout
Add a new configuration option “imapidletimeout” which, if
greater than zero, specifies an upper limit in seconds for idle
connections. The value defaults to zero (not set).

RFC 2177 recommends that a client re-issue the IDLE command at
least every 29 minutes if it wishes to continue, otherwise the
server is free to treat the client as disconnected.

The rationale is that sometimes connections aren’t properly
reset. Currently, a connection is not collected if it was in IDLE
state at that point. If this happens repeatedly, imapd keeps
accumulating dead connections which can cause DOS. This patch
solves the problem by forcing imapd to stop idling after
exceeding the configured timeout.