imapparse: don't unget after reading a bad literal
We only reject literals that contain NULs, and only in certain
cases.  But in these cases, if we put back the last byte, and
that byte happened to be a \n, the subsequent eatline() would
lose sync again.

Better to consume the whole literal regardless of whether we
accept it or not.