tools/config2rst: fix regex bug and lies
The "we explicitly do not match deprecated options" comment was
misleading:  even though the regex didn't expect or capture a $6
field, the unbounded wildcards in the other fields allowed
deprecated options to still be matched, misparsed and output anyway.

And we need to parse and output deprecated options, so we can
document that we're deprecated!  We've been getting away with
the bad parsing because we weren't doing anything with the
extra fields anyway.

So this hardens up the matching to make sure the values are parsed
into the correct fields, explicitly matches deprecated options too,
and removes the lie that we don't.